requirements.txt
aniso8601==10.0.0
blinker==1.9.0
click==8.1.8
Flask==3.1.0
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
packaging==24.2
pluggy==1.5.0
psycopg2==2.9.10
pytest==8.3.4
python-dotenv==1.0.1
pytz==2024.2
six==1.17.0
SQLAlchemy==2.0.37
typing_extensions==4.12.2
Werkzeug==3.1.3
